Do a better job placing #line pragmas.
The problem with the SrcLoc constructor is that it may occur just after
a line. Now when we need to output a line, we wait to output the line
until we have to output something more. At this point we really do need
to output the pending line, but we may also have more up-to-date
location information.